+++ src/lib.rs
@@ -1,130 +1,9 @@
 use pyo3::prelude::{pymodule, PyModule, PyResult, Python};
-use rand::prelude::*;
-use rayon::prelude::*;
 use numpy::{PyArray, PyArray1, ToPyArray};
-use rug::{float::Round, Float};
 
-
-fn uniform(scale: f64) -> f64 {
-    /// Returns a sample from the [0, scale) uniform distribution
-    ///
-
-    let mut rng = rand::thread_rng();
-    rng.gen::<f64>()
-}
-
-
-fn exponential(scale: f64) -> f64 {
-    /// Returns a sample from the exponential distribution
-    ///
-    /// # Arguments
-    ///
-    /// * `scale` - The scale parameter of the exponential distribution
-
-    let sample = -scale * uniform(1.0).ln();
-    sample
-}
-
-
-fn laplace(scale: f64) -> f64 {
-    /// Returns one sample from the Laplace distribution
-    ///
-    /// # Arguments
-    ///
-    /// * `scale` - The scale parameter of the Laplace distribution
-
-    let y = uniform(1.0) - 0.5;
-    let sgn = y.signum();
-    sgn * (2.0 * sgn * y).ln() * scale
-}
-
-
-fn geometric(scale: f64) -> f64 {
-    /// Returns a sample from the geometric distribution
-    ///
-    /// # Arguments
-    ///
-    /// * `scale` - The scale parameter of the geometric distribution
-
-    (uniform(1.0).ln() / (1.0 - scale).ln()).floor()
-}
-
-
-fn two_sided_geometric(scale: f64) -> f64 {
-    /// Returns a sample from the two sided geometric distribution
-    ///
-    /// # Arguments
-    ///
-    /// * `scale` - The scale parameter of the two sided geometric distribution
-
-    let y = (uniform(1.0) - 0.5) * (1.0 + scale);
-    let sgn = y.signum();
-    sgn * ((sgn * y).ln() / scale.ln()).floor()
-}
-
-
-fn double_uniform(scale: f64) -> f64 {
-    /// Returns a sample from the [0, scale) uniform distribution
-    ///
-
-    let mut rng = rand::thread_rng();
-    let exponent: f64 = geometric(0.5) + 53.0;
-    let mut significand = (rng.gen::<u64>() >> 11) | (1 << 52);
-    scale * (significand as f64) * 2.0_f64.powf(-exponent)
-}
-
-
-fn vectorize(scale: f64, num: usize, func: fn(f64) -> f64) -> Vec<f64> {
-    /// Vectorize a distribution sampler
-    ///
-    /// # Arguments
-    ///
-    /// * `scale` - The scale parameter of the distribution
-    /// * `num` - The number of samples to draw
-    /// * `func` - The distribution function
-
-    let mut samples: Vec<f64> = vec![0.0; num];
-    samples.par_iter_mut().for_each(|p| *p = func(scale));
-    samples
-}
-
-
-fn all_above_threshold(
-    data: Vec<f64>, scale: f64, threshold: f64
-) -> Vec<usize>{
-    data.par_iter().positions(|&p| p + laplace(scale) > threshold).collect()
-}
-
-
-fn clamp(x: f64, bound: f64) -> f64 {
-    if x < -bound {
-        -bound
-    } else if x > bound {
-        bound
-    } else {
-        x
-    }
-}
-
-
-fn ln_rn(x: f64) -> f64 {
-    let mut y = Float::with_val(53, x);
-    let dir = y.ln_round(Round::Nearest);
-    y.to_f64()
-}
-
-
-fn snapping(
-    data: Vec<f64>, bound: f64, lambda: f64, quanta: f64
-) -> Vec<f64> {
-    data.par_iter()
-        .map(|&p| clamp(p, bound))
-        .map(|p| p + lambda * ln_rn(double_uniform(1.0)) * (uniform(1.0) - 0.5).signum())
-        .map(|p| quanta * (p / quanta).round())
-        .map(|p| clamp(p, bound))
-        .collect()
-}
-
+mod utils;
+mod samplers;
+mod mechanisms;
 
 ///// A Python module implemented in Rust.
 ///// Exports the rust functions to python.
@@ -136,7 +15,7 @@ fn backend(py: Python, m: &PyModule) -> PyResult<()> {
         /// Simple python wrapper of the exponential function. Converts
         /// the rust vector into a numpy array
 
-        vectorize(1.0, num, uniform).to_pyarray(py)
+        utils::vectorize(1.0, num, samplers::uniform).to_pyarray(py)
     }
 
     #[pyfn(m, "exponential")]
@@ -144,7 +23,7 @@ fn backend(py: Python, m: &PyModule) -> PyResult<()> {
         /// Simple python wrapper of the exponential function. Converts
         /// the rust vector into a numpy array
 
-        vectorize(scale, num, exponential).to_pyarray(py)
+        utils::vectorize(scale, num, samplers::exponential).to_pyarray(py)
     }
 
     #[pyfn(m, "laplace")]
@@ -152,7 +31,7 @@ fn backend(py: Python, m: &PyModule) -> PyResult<()> {
         /// Simple python wrapper of the laplace function. Converts
         /// the rust vector into a numpy array
 
-        vectorize(scale, num, laplace).to_pyarray(py)
+        utils::vectorize(scale, num, samplers::laplace).to_pyarray(py)
     }
 
     #[pyfn(m, "geometric")]
@@ -160,7 +39,7 @@ fn backend(py: Python, m: &PyModule) -> PyResult<()> {
         /// Simple python wrapper of the geometric function. Converts
         /// the rust vector into a numpy array
 
-        vectorize(scale, num, geometric).to_pyarray(py)
+        utils::vectorize(scale, num, samplers::geometric).to_pyarray(py)
     }
 
     #[pyfn(m, "two_sided_geometric")]
@@ -168,14 +47,14 @@ fn backend(py: Python, m: &PyModule) -> PyResult<()> {
         /// Simple python wrapper of the two sided geometric function. Converts
         /// the rust vector into a numpy array
 
-        vectorize(scale, num, two_sided_geometric).to_pyarray(py)
+        utils::vectorize(scale, num, samplers::two_sided_geometric).to_pyarray(py)
     }
 
     #[pyfn(m, "double_uniform")]
     fn py_double_uniform(py: Python, num: usize) -> &PyArray1<f64>{
         /// Simple python wrapper of the exponential function. Converts
         /// the rust vector into a numpy array
-        vectorize(1.0, num, double_uniform).to_pyarray(py)
+        utils::vectorize(1.0, num, samplers::double_uniform).to_pyarray(py)
     }
 
     #[pyfn(m, "all_above_threshold")]
@@ -186,7 +65,7 @@ fn backend(py: Python, m: &PyModule) -> PyResult<()> {
         /// Simple python wrapper of the exponential function. Converts
         /// the rust vector into a numpy array
         let data = data.to_vec().unwrap();
-        all_above_threshold(data, scale, threshold).to_pyarray(py)
+        mechanisms::all_above_threshold(data, scale, threshold).to_pyarray(py)
     }
 
     #[pyfn(m, "snapping")]
@@ -197,14 +76,14 @@ fn backend(py: Python, m: &PyModule) -> PyResult<()> {
         /// Simple python wrapper of the exponential function. Converts
         /// the rust vector into a numpy array
         let data = data.to_vec().unwrap();
-        snapping(data, bound, lambda, quanta).to_pyarray(py)
+        mechanisms::snapping(data, bound, lambda, quanta).to_pyarray(py)
     }
 
     #[pyfn(m, "ln_rn")]
     fn py_ln_rn(x: f64) -> f64 {
         /// Simple python wrapper of the exponential function. Converts
         /// the rust vector into a numpy array
-        ln_rn(x)
+        utils::ln_rn(x)
     }
 
     Ok(())
